backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-jazzmin>=3.0,<4.0
python-dotenv>=1.0,<2.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.6.0,<5.0
pytest>=8.0,<9.0
pytest-django>=4.5,<5.0
django-allauth>=0.57,<0.62
dj-rest-auth[with_social]>=6.0,<7.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
python-json-logger>=2.0,<3.0
requests>=2.31,<3.0
# Sentry/GlitchTip (opcional) - Descomente se quiser usar Sentry ou GlitchTip para logging
# GlitchTip é compatível com a API do Sentry, então usa os mesmos SDKs
# sentry-sdk[django]>=2.0,<3.0

